requirements.txt
pyqt6>=6.0.0
pillow>=9.0.0
pyperclip>=1.8.0
pyautogui>=0.9.0
pyobjc>=8.0
pytest>=7.0.0
rumps>=0.4.0
pynput>=1.8.0